fcoei: overflow in conversion
In file included from ../../common/io/fibre-channel/fca/fcoei/fcoei_lv.c:61:
../../common/io/fibre-channel/fca/fcoei/fcoei_lv.c: In function 'fcoei_fill_els_
prli_cmd':
../../common/sys/fcoe/fcoe_common.h:201:41: error: overflow in conversion from '
int' to 'uint8_t' {aka 'unsigned char'} changes value from '(((int)fcp_spp->orig
_process_assoc_valid << 15 | (int)fcp_spp->resp_process_assoc_valid << 14) | (in
t)fcp_spp->establish_image_pair << 13) & 255' to '0' [-Werror=overflow]
  201 |                 ((uint8_t *)(x_b))[1] = 0xFF & (x_v);           \
      |                                         ^~~~
../../common/io/fibre-channel/fca/fcoei/fcoei_lv.c:1459:9: note: in expansion of
 macro 'FCOE_V2B_2'
 1459 |         FCOE_V2B_2(((fcp_spp->orig_process_assoc_valid << 15) |
      |         ^~~~~~~~~~
cc1: all warnings being treated as errors

+++ usr/src/uts/common/sys/fcoe/fcoe_common.h
@@ -198,7 +198,7 @@ typedef struct fcoe_client {
 
 #define	FCOE_V2B_2(x_v, x_b)					\
 	{							\
-		((uint8_t *)(x_b))[1] = 0xFF & (x_v);		\
+		((uint8_t *)(x_b))[1] = (uint8_t)(0xFF & (x_v)); \
 		((uint8_t *)(x_b))[0] = 0xFF & ((x_v) >> 8);	\
 	}
